+++ api-client/lib/Api/SearchApi.php
@@ -27,6 +27,7 @@
 
 namespace Elio\ElioBatteryIncludedApiClient\Api;
 
+use Elio\ElioDataDiscovery\Core\Util\PropertyUtil;
 use Elio\ElioDataDiscovery\Swagger\ClientApiException;
 use Elio\ElioDataDiscovery\Swagger\ClientConfiguration;
 use Elio\ElioDataDiscovery\Swagger\ClientHeaderSelector;
@@ -265,6 +266,8 @@ class SearchApi
             foreach ($filters as $key => $filter) {
                 if (is_array($filter)) {
                     foreach ($filter as $item) {
+                        $item = PropertyUtil::decodeStringFromUnescapedUnicode($item);
+                        $key = PropertyUtil::decodeStringFromUnescapedUnicode($key);
                         $queryParams[$key][] = ClientObjectSerializer::toQueryValue($item, null);
                     }
                 } else {
@@ -355,16 +358,16 @@ class SearchApi
      * Suggest
      *
      * @param string $q q (optional)
-     * @param string $type type (optional)
+     * @param array $filters filters (optional)
      * @param string $x_bi_api_key x_bi_api_key (optional)
      *
      * @throws ClientApiException on non-2xx response
      * @throws InvalidArgumentException
      *  TODO: Add request into parameters
      */
-    public function suggest($q, $language, $type = null, $x_bi_api_key = null)
+    public function suggest($q, $language, $filters = null, $x_bi_api_key = null)
     {
-        list($response) = $this->suggestWithHttpInfo($q, $language, $type, $x_bi_api_key);
+        list($response) = $this->suggestWithHttpInfo($q, $language, $filters, $x_bi_api_key);
         return $response;
     }
 
@@ -374,17 +377,17 @@ class SearchApi
      * Suggest
      *
      * @param string $q (optional)
-     * @param string $type type (optional)
+     * @param array $filters filters (optional)
      * @param string $x_bi_api_key (optional)
      *
      * @return array of null, HTTP status code, HTTP response headers (array of strings)
      * @throws InvalidArgumentException
      * @throws ApiException on non-2xx response
      */
-    public function suggestWithHttpInfo($q, $language, $type = null, $x_bi_api_key = null)
+    public function suggestWithHttpInfo($q, $language, $filters = null, $x_bi_api_key = null)
     {
         $returnType = '\Elio\ElioBatteryIncludedApiClient\Model\SuggestionResult[]';
-        $request = $this->suggestRequest($q, $language, $type, $x_bi_api_key);
+        $request = $this->suggestRequest($q, $language, $filters, $x_bi_api_key);
         try {
             $options = $this->createHttpClientOption();
             try {
@@ -441,15 +444,15 @@ class SearchApi
      * Suggest
      *
      * @param string $q (optional)
-     * @param string $type type (optional)
+     * @param array $filters filters (optional)
      * @param string $x_bi_api_key (optional)
      *
      * @return PromiseInterface
      * @throws InvalidArgumentException
      */
-    public function suggestAsync($q, $language, $type = null, $x_bi_api_key = null)
+    public function suggestAsync($q, $language, $filters = null, $x_bi_api_key = null)
     {
-        return $this->suggestAsyncWithHttpInfo($q, $language, $type, $x_bi_api_key)
+        return $this->suggestAsyncWithHttpInfo($q, $language, $filters, $x_bi_api_key)
             ->then(
                 function ($response) {
                     return $response[0];
@@ -463,16 +466,16 @@ class SearchApi
      * Suggest
      *
      * @param string $q (optional)
-     * @param string $type type (optional)
+     * @param array $filters filters (optional)
      * @param string $x_bi_api_key (optional)
      *
      * @return PromiseInterface
      * @throws InvalidArgumentException
      */
-    public function suggestAsyncWithHttpInfo($q, $language, $type = null, $x_bi_api_key = null)
+    public function suggestAsyncWithHttpInfo($q, $language, $filters = null, $x_bi_api_key = null)
     {
         $returnType = '';
-        $request = $this->suggestRequest($q, $language, $type, $x_bi_api_key);
+        $request = $this->suggestRequest($q, $language, $filters, $x_bi_api_key);
 
         return $this->client
             ->sendAsync($request, $this->createHttpClientOption())
@@ -501,13 +504,13 @@ class SearchApi
      * Create request for operation 'suggest'
      *
      * @param string $q (optional)
-     * @param string $type type (optional)
+     * @param array $filters filters (optional)
      * @param string $x_bi_api_key (optional)
      *
      * @return Request
      * @throws InvalidArgumentException
      */
-    protected function suggestRequest($q, $language, $type = null, $x_bi_api_key = null)
+    protected function suggestRequest($q, $language, $filters = null, $x_bi_api_key = null)
     {
         $resourcePath = sprintf('/api/v1/collections/%s/documents/suggest', $this->config->getCollection());
         $formParams = [];
@@ -523,8 +526,19 @@ class SearchApi
 
         $queryParams['v[locale]'] = ClientObjectSerializer::toQueryValue($language, null);
 
-        if ($type !== null) {
-            $queryParams['f[type]'] = ClientObjectSerializer::toQueryValue($type, null);
+        if (is_array($filters)) {
+            foreach ($filters as $key => $filter) {
+                if (is_array($filter)) {
+                    foreach ($filter as $item) {
+                        $queryParams[$key][] = ClientObjectSerializer::toQueryValue($item, null);
+                    }
+                } else {
+                    $queryParams[$key][] = ClientObjectSerializer::toQueryValue($filter, null);
+                }
+
+            }
+        } else {
+            $queryParams['filters'] = ClientObjectSerializer::toQueryValue($filters, null);
         }
 
         // header params