+++ src/Api/Search/SearchApiDecorator.php
@@ -44,7 +44,7 @@ use Elio\ElioDataDiscovery\Api\Search\Request\ProductSearchRequest;
 use Elio\ElioDataDiscovery\Api\Search\Request\SearchRequest;
 use Elio\ElioDataDiscovery\Api\Search\SearchApi;
 use Elio\ElioDataDiscovery\Api\Transform\Transformer;
-use Elio\ElioDataDiscovery\Configuration\ElioDataDiscoveryConfigService;
+use Elio\ElioDataDiscovery\Configuration\ElioDataDiscoveryConfigServiceInterface;
 use Elio\ElioDataDiscovery\Core\Exception\InvalidTypeException;
 use Elio\ElioDataDiscovery\Core\FilterRestrictions\FilterEntity;
 use Elio\ElioDataDiscovery\Core\Logging\RequestLoggingService;
@@ -81,7 +81,7 @@ class SearchApiDecorator extends SearchApi
         private readonly SystemConfigService $systemConfigService,
         private readonly EntityRepository $filterRepository,
         private readonly RequestLoggingService $requestLoggingService,
-        private readonly ElioDataDiscoveryConfigService $configService
+        private readonly ElioDataDiscoveryConfigServiceInterface $configService
     ) {
         parent::__construct($logger);
     }